refactor: split dumper and restorer
